+++ yamcs-xtce/src/main/java/org/yamcs/xtce/FloatValidRange.java
@@ -1,11 +1,13 @@
 package org.yamcs.xtce;
 
+import org.yamcs.utils.DoubleRange;
+
 /**
  * Holds the min/max of a ValidRange for floating point.
  * @author nm
  *
  */
-public class FloatValidRange extends FloatRange {
+public class FloatValidRange extends DoubleRange {
 	/**
 	 * The range specified applies to the destination data type if true, 
 	 * or the raw source data type if false. The default is false and reflects the more likely scenario of checking raw values before conversion or calibration in telemetry.
