+++ yamcs-xtce/src/main/java/org/yamcs/xtce/IntegerParameterType.java
@@ -7,6 +7,8 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 
+import org.yamcs.utils.DoubleRange;
+
 public class IntegerParameterType extends IntegerDataType implements ParameterType {
     private static final long serialVersionUID = 2L;
     List<NumericContextAlarm> contextAlarmList = null;
@@ -39,23 +41,23 @@ public class IntegerParameterType extends IntegerDataType implements ParameterTy
         return alarm.getStaticAlarmRanges();
     }
 
-    public void setDefaultWatchAlarmRange(FloatRange watchRange) {
+    public void setDefaultWatchAlarmRange(DoubleRange watchRange) {
         getAlarmRanges(null).watchRange=watchRange;
     }
 
-    public void setDefaultWarningAlarmRange(FloatRange warningRange) {
+    public void setDefaultWarningAlarmRange(DoubleRange warningRange) {
         getAlarmRanges(null).warningRange=warningRange;
     }
 
-    public void setDefaultDistressAlarmRange(FloatRange distressRange) {
+    public void setDefaultDistressAlarmRange(DoubleRange distressRange) {
         getAlarmRanges(null).distressRange=distressRange;
     }
 
-    public void setDefaultCriticalAlarmRange(FloatRange criticalRange) {
+    public void setDefaultCriticalAlarmRange(DoubleRange criticalRange) {
         getAlarmRanges(null).criticalRange=criticalRange;
     }
 
-    public void setDefaultSevereAlarmRange(FloatRange severeRange) {
+    public void setDefaultSevereAlarmRange(DoubleRange severeRange) {
         getAlarmRanges(null).severeRange=severeRange;
     }
 
@@ -63,14 +65,14 @@ public class IntegerParameterType extends IntegerDataType implements ParameterTy
      * Adds a new, or unions with an existing range for the specified context and level
      * @param contextMatch use <tt>null</tt> for the default context
      */
-    public void addAlarmRange(MatchCriteria contextMatch, FloatRange range, AlarmLevels level) {
+    public void addAlarmRange(MatchCriteria contextMatch, DoubleRange range, AlarmLevels level) {
         getAlarmRanges(contextMatch).addRange(range, level);
     }
     /**
      * Adds a new, or unions with an existing watch range for the specified context
      * @param contextMatch use <tt>null</tt> for the default context
      */
-    public void addWatchAlarmRange(MatchCriteria contextMatch, FloatRange watchRange) {
+    public void addWatchAlarmRange(MatchCriteria contextMatch, DoubleRange watchRange) {
         getAlarmRanges(contextMatch).addWatchRange(watchRange);
     }
 
@@ -78,7 +80,7 @@ public class IntegerParameterType extends IntegerDataType implements ParameterTy
      * Adds a new, or unions with an existing warning range for the specified context
      * @param contextMatch use <tt>null</tt> for the default context
      */
-    public void addWarningAlarmRange(MatchCriteria contextMatch, FloatRange warningRange) {
+    public void addWarningAlarmRange(MatchCriteria contextMatch, DoubleRange warningRange) {
         getAlarmRanges(contextMatch).addWarningRange(warningRange);
     }
 
@@ -86,7 +88,7 @@ public class IntegerParameterType extends IntegerDataType implements ParameterTy
      * Adds a new, or unions with an existing distress range for the specified context
      * @param contextMatch use <tt>null</tt> for the default context
      */
-    public void addDistressAlarmRange(MatchCriteria contextMatch, FloatRange distressRange) {
+    public void addDistressAlarmRange(MatchCriteria contextMatch, DoubleRange distressRange) {
         getAlarmRanges(contextMatch).addDistressRange(distressRange);
     }
 
@@ -94,7 +96,7 @@ public class IntegerParameterType extends IntegerDataType implements ParameterTy
      * Adds a new, or unions with an existing critical range for the specified context
      * @param contextMatch use <tt>null</tt> for the default context
      */
-    public void addCriticalAlarmRange(MatchCriteria contextMatch, FloatRange criticalRange) {
+    public void addCriticalAlarmRange(MatchCriteria contextMatch, DoubleRange criticalRange) {
         getAlarmRanges(contextMatch).addCriticalRange(criticalRange);
     }
 
@@ -102,7 +104,7 @@ public class IntegerParameterType extends IntegerDataType implements ParameterTy
      * Adds a new, or unions with an existing severe range for the specified context
      * @param contextMatch use <tt>null</tt> for the default context
      */
-    public void addSevereAlarmRange(MatchCriteria contextMatch, FloatRange severeRange) {
+    public void addSevereAlarmRange(MatchCriteria contextMatch, DoubleRange severeRange) {
         getAlarmRanges(contextMatch).addSevereRange(severeRange);
     }
